refactor: move the last of root.c elsewhere
