+++ Makefile.am
@@ -69,7 +69,6 @@ watchman_SOURCES = \
 	listener-user.c   \
 	clientmode.c \
 	main.c       \
-	root.c       \
 	root/ageout.c       \
 	root/crawler.c       \
 	root/dir.c       \
