--- root.c
@@ -1,55 +1,0 @@
-/* Copyright 2012-present Facebook, Inc.
- * Licensed under the Apache License, Version 2.0 */
-
-#include "watchman.h"
-
-void free_file_node(w_root_t *root, struct watchman_file *file)
-{
-  root->watcher_ops->file_free(file);
-  if (file->symlink_target) {
-    w_string_delref(file->symlink_target);
-  }
-  free(file);
-}
-
-void w_root_addref(w_root_t *root)
-{
-  w_refcnt_add(&root->refcnt);
-}
-
-void w_root_schedule_recrawl(w_root_t *root, const char *why)
-{
-  if (!root->should_recrawl) {
-    if (root->last_recrawl_reason) {
-      w_string_delref(root->last_recrawl_reason);
-    }
-
-    root->last_recrawl_reason = w_string_make_printf(
-        "%.*s: %s",
-        root->root_path->len, root->root_path->buf, why);
-
-    w_log(W_LOG_ERR, "%.*s: %s: scheduling a tree recrawl\n",
-        root->root_path->len, root->root_path->buf, why);
-  }
-  root->should_recrawl = true;
-  signal_root_threads(root);
-}
-
-// Caller must have locked root
-json_t *w_root_trigger_list_to_json(struct read_locked_watchman_root *lock)
-{
-  w_ht_iter_t iter;
-  json_t *arr;
-
-  arr = json_array();
-  if (w_ht_first(lock->root->commands, &iter)) do {
-    struct watchman_trigger_command *cmd = w_ht_val_ptr(iter.value);
-
-    json_array_append(arr, cmd->definition);
-  } while (w_ht_next(lock->root->commands, &iter));
-
-  return arr;
-}
-
-/* vim:ts=2:sw=2:et:
- */
