+++ root/init.c
@@ -198,6 +198,10 @@ void w_root_teardown(w_root_t *root) {
   }
 }
 
+void w_root_addref(w_root_t *root) {
+  w_refcnt_add(&root->refcnt);
+}
+
 void w_root_delref(w_root_t *root) {
   if (!w_refcnt_del(&root->refcnt)) return;
